ISSCC Session 34. Computing In Memory
Welcome to Session 34 of the prestigious International Solid-State Circuits Conference, where today we will be exploring the frontiers of "In-Memory Computing." As we gather the brightest minds in the field, we are excited to delve into a session that promises to be both enlightening and transformative for the future of integrated circuit design and computing technology.

Session Overview:
In this session, we have carefully curated a lineup of papers that will take us on a journey through the latest advancements in in-memory computing. Our speakers will present innovative research and breakthroughs that address the pressing need for higher computational efficiency and energy optimization in the era of big data and artificial intelligence.