The advent of non-volatile memory (NVM) technology
The advent of non-volatile memory (NVM) technologies has revolutionized the landscape of data storage and processing, offering a tantalizing blend of high density, low power consumption, and persistence of data even when power is removed. This special collection of papers, featured in our esteemed publication, delves into the burgeoning field of In-Memory Computing (IMC) architectures that leverage the unique attributes of non-volatile memory.